deployment fixes
