--- .unotes/templates/title_date.hbs
@@ -1,4 +1,0 @@
-# {{capitalizeAll title}}
-
-{{formatDate date "llll"}}
-
