--- .unotes/unotes_meta.json
@@ -1,75 +1,0 @@
-{
-  "name": "",
-  "isOrdered": true,
-  "folders": {
-    "out": {
-      "name": "out",
-      "isOrdered": true,
-      "folders": {
-        "components": {
-          "name": "components",
-          "isOrdered": true,
-          "folders": {},
-          "files": {}
-        },
-        "test": {
-          "name": "test",
-          "isOrdered": true,
-          "folders": {},
-          "files": {}
-        }
-      },
-      "files": {}
-    },
-    "src": {
-      "name": "src",
-      "isOrdered": true,
-      "folders": {
-        "components": {
-          "name": "components",
-          "isOrdered": true,
-          "folders": {
-            "start": {
-              "name": "start",
-              "isOrdered": true,
-              "folders": {},
-              "files": {}
-            }
-          },
-          "files": {}
-        },
-        "fonts": {
-          "name": "fonts",
-          "isOrdered": true,
-          "folders": {},
-          "files": {}
-        },
-        "style": {
-          "name": "style",
-          "isOrdered": true,
-          "folders": {},
-          "files": {}
-        },
-        "test": {
-          "name": "test",
-          "isOrdered": true,
-          "folders": {
-            "suite": {
-              "name": "suite",
-              "isOrdered": true,
-              "folders": {},
-              "files": {}
-            }
-          },
-          "files": {}
-        }
-      },
-      "files": {}
-    }
-  },
-  "files": {
-    "README": 0,
-    "CHANGELOG": 1,
-    "vsc-extension-quickstart": 2
-  }
-}
