+++ package.json
@@ -1,9 +1,9 @@
 {
 	"name": "entry",
 	"displayName": "entry",
-	"publisher": "Entry",
+	"publisher": "entry",
 	"description": "Write notes for your code",
-	"version": "0.0.1",
+	"version": "0.0.92",
 	"engines": {
 		"vscode": "^1.83.0"
 	},
@@ -63,19 +63,20 @@
 		]
 	},
 	"scripts": {
-		"vscode:prepublish": "npm run compile",
+		"vscode:prepublish": "npm run compile && npm run copy-assets",
 		"compile": "tsc -p ./",
 		"watch": "tsc -watch -p ./",
 		"pretest": "npm run compile && npm run lint",
 		"lint": "eslint src --ext ts",
-		"test": "node ./out/test/runTest.js"
+		"test": "node ./out/test/runTest.js",
+		"copy-assets": "xcopy /E /I src\\style\\* out\\style && xcopy /E /I src\\utils\\* out\\utils"
 	},
 	"devDependencies": {
 		"@types/fs-extra": "^11.0.4",
 		"@types/mocha": "^10.0.3",
 		"@types/node": "18.x",
 		"@types/quill": "^2.0.14",
-		"@types/vscode": "^1.84.0",
+		"@types/vscode": "^1.83.0",
 		"@typescript-eslint/eslint-plugin": "^6.9.0",
 		"@typescript-eslint/parser": "^6.9.0",
 		"@vscode/test-electron": "^2.3.6",
