+++ src/components/note/note.ts
@@ -1,29 +1,20 @@
 import * as vscode from "vscode";
 import { loadFile } from "../../utils/saveFile";
+import { header } from "../../utils/header";
 
 export async function getWebviewNote(webview: vscode.Webview, context: any, fileName: string, filePath: string, currentPage?: string) {
-	const codiconsUri = webview.asWebviewUri(vscode.Uri.joinPath(context.extensionUri, "node_modules", "@vscode/codicons", "dist", "codicon.css"));
-	const generalStyles = webview.asWebviewUri(vscode.Uri.joinPath(context.extensionUri, "src/style", "general.css"));
-	const script = webview.asWebviewUri(vscode.Uri.joinPath(context.extensionUri, "src/utils", "script.js"));
+	const script = webview.asWebviewUri(vscode.Uri.joinPath(context.extensionUri, "out/utils", "script.js"));
 
 	const isDark = vscode.window.activeColorTheme?.kind === vscode.ColorThemeKind.Dark;
 
 	const loadedContent = await loadFile(fileName, filePath, context);
 
+	const htmlHeader = await header(webview, context);
+
 	return `<!DOCTYPE html>
 	<html lang="en">
-		<head>
-			<meta charset="UTF-8" />
-			<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-			<script href="highlight.js"></script>
-			<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/styles/default.min.css">
-			<script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/highlight.min.js"></script>
-			<script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/languages/go.min.js"></script>
-			<link href="https://cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet">
-			<script src="https://cdn.quilljs.com/1.3.6/quill.js"></script>
-			<link rel="stylesheet" href="${codiconsUri}">
-			<link rel="stylesheet" href="${generalStyles}">
-		</head>
+		${htmlHeader}
+
 		<body>
 			<div class="toolbar-container">
 			<div class="back-button">
