+++ src/utils/scriptImport.ts
@@ -1,0 +1,14 @@
+import * as vscode from "vscode";
+
+export async function scriptImport(webview: vscode.Webview, context: any) {
+    const script = webview.asWebviewUri(vscode.Uri.joinPath(context.extensionUri, "out/utils", "script.js"));
+
+    const isDark = vscode.window.activeColorTheme?.kind === vscode.ColorThemeKind.Dark;
+
+    return `
+        <script src="${script}"></script>
+        <script>
+	        updateTheme(${isDark});
+        </script>
+    `
+}
